refactor(rule_based):change lane_change_assistant vairables
