+++ simulation/main.py
@@ -3,16 +3,3 @@ import numpy as np
 if __name__ == '__main__':
     g = Game()
     g.run()
-    # lane_points = None
-    # distance_points = None
-    # for road in range(len(g.map1.roads)):
-    #     print (g.map1.roads[road].length)
-    #     for lane in range(len(g.map1.roads[road].lanes)):
-    #         lane_points = g.map1.roads[road].lanes[lane].lane_points
-    #         print (lane_points)
-    #         print(len(lane_points))
-    #         distance_points = g.map1.roads[road].lanes[lane].distance_points
-    #         break
-    #     print()
-    #     break
-
