+++ simulation/traffic/vehicle/vehicle_class.py
